Module Idl.Exn

type rpcfn = Rpc.call -> Rpc.response

type client_implementation

module GenClient (R : sig ... end) : sig ... end

This module generates exception-raising Client modules from RPC declarations. See the GenClient module for a description of the common entries.

type server_implementation

val server : server_implementation -> rpcfn

val combine : server_implementation list -> server_implementation

module GenServer () : sig ... end

Generates a server, like GenServer, but for an implementation that raises exceptions instead of returning a result.