Module Internals

val encode : (char -> string option) -> string -> string

Encodes a string using the given translation function that maps a character to a string that is its encoded version, if that character needs encoding.