Module Idl.Make

Parameters

M : MONAD

Signature

type client_implementation

type server_implementation

module type RPCTRANSFORMER = sig ... end

module T : RPCTRANSFORMER

The module !T, the RPC MONAD transformer, defines the minimal set of types and functions needed for the !GenClient and !GenServer modules to generate clients and servers. These allow to provide different syncronous and asynctronous engines for the client and server implementations.

module ErrM : sig ... end

!ErrM defines monad to use for the implementation and combination of RPC functions

module GenClient : functor () sig ... end

This module generates Client modules from RPC declarations.

module GenServer : functor () sig ... end

This module generates a server that dispatches RPC calls to their implementations.

val server : server_implementation -> T.rpcfn

val combine : server_implementation list -> server_implementation